2011-02-14 14:39 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * contrib/hbwin/hbwin.hbx   * contrib/hbwin/win_regc.c   * contrib/hbwin/win_reg.prg   * contrib/hbwin/tests/testreg.prg     + Added WIN_REGDELETE() high-level function.     + Added WIN_REGDELETEKEY() Windows API (non-1-to-1) wrapper.     + Added WIN_REGDELETEVALUE() Windows API (non-1-to-1) wrapper.
  * contrib/hbqt/hbmk2_qt.hbs
  * contrib/hbqt/qtgui/hbqtgui.h
  * contrib/hbqt/qtcore/hbqtcore.h
    + Cast HBQT_TYPE_ constants to unsigned int.

+++ harbour/contrib/hbqt/qtcore/hbqtcore.h
@@ -23,6 +23,6 @@ HB_EXTERN_END
 #define hbqt_par_HBQSlots( n )                                  ( ( HBQSlots                                    * ) hbqt_gcpointer( n ) )
 #define hbqt_par_HBQString( n )                                 ( ( HBQString                                   * ) hbqt_gcpointer( n ) )
 
-#define HBQT_TYPE_QSize                                         0x5E28750F
+#define HBQT_TYPE_QSize                                         ( ( unsigned int ) 0x5E28750F )
 
 #endif /* __HBQTCORE_H */
